+++ libs/common/src/lib/types/toggle-option.type.ts
@@ -1,0 +1,9 @@
+interface BaseToggleOption {
+  value: string;
+}
+
+export type ToggleOption = BaseToggleOption &
+  (
+    | { iconName: string; label?: never; title: string }
+    | { iconName?: never; label: string; title?: never }
+  );